+++ CHANGELOG.md
@@ -9,6 +9,7 @@ To get the diff between two versions, go to https://github.com/forxer/tao/compar
 	- search tools enhancements
 	- add support service provider
 	- register more default services provider
+	- add X-Frame-Options Response Headers
 - 0.8.1
 	- add filesystem service provider
 	- add finder service provider