+++ CHANGELOG.md
@@ -7,10 +7,12 @@ To get the diff between two versions, go to https://github.com/forxer/tao/compar
 
 - 0.8.3.dev
     - bug #3 off the class attribute of the previous attribute
+    - small code quality enhancements
 - 0.8.2
     - search tools enhancements
     - add support service provider
     - register more default services provider
+    - add X-Frame-Options Response Headers
 - 0.8.1
     - add filesystem service provider
     - add finder service provider