+++ CHANGELOG.md
@@ -5,6 +5,8 @@ This changelog does not reference all the changes, only the most relevant to be 
 To get the diff for a specific change, go to https://github.com/forxer/tao/commit/XXX where XXX is the change hash
 To get the diff between two versions, go to https://github.com/forxer/tao/compare/0.6...0.7
 
+- 0.8.4.dev
+    - add some core triggers
 - 0.8.3
     - bug #3 off the class attribute of the previous attribute
     - update Symfony Components to 2.6.*